Fixing .gitignore
There was an error in 8c924b198378bbfecea0c56585ac852c538daec3 -- looks like the intent was to add the `dist/` directory to the gitignore but it got added as `dist` (which only ignores files called "dist" but not directories). There is a similar error on `node_modules/` but it didn't matter because the recursive `**node_modules/` patten caught the base case anyway.