Converting docs to Github markdown
The code in the docs doesn't highlight in Github, and on inspection it's because the code is wrapped in `pre` tags instead of using Github's code/syntax highlighting. I updated all the `pre` blocks and converted most URLs to use markdown style formatting (with the exception of the image wrapped in a link, since that still requires raw HTML).